Add missing policies
